+++ MercadoPagoSDK/MercadoPagoSDK/Config/PXAdvancedConfiguration.swift
@@ -15,5 +15,5 @@ open class PXAdvancedConfiguration: NSObject {
     open var binaryMode: Bool = false
     open var bankDealsEnabled: Bool = true
     open var reviewConfirmConfiguration: PXReviewConfirmConfiguration = PXReviewConfirmConfiguration()
-    open var paymentResultScreenPreference: PaymentResultScreenPreference = PaymentResultScreenPreference()
+    open var paymentResultConfiguration: PXPaymentResultConfiguration = PXPaymentResultConfiguration()
 }
