--- MercadoPagoSDK/MercadoPagoSDK/Config/PXResultScreenUIProtocol.swift
@@ -1,14 +1,0 @@
-//
-//  PXResultScreenUIProtocol.swift
-//  MercadoPagoSDK
-//
-//  Created by Juan sebastian Sanzone on 6/8/18.
-//  Copyright © 2018 MercadoPago. All rights reserved.
-//
-
-import Foundation
-
-/** :nodoc: - Not intended for public use. */
-public protocol PXResultScreenUIProtocol {
-
-}
