+++ MercadoPagoSDK/MercadoPagoSDK/MercadoPagoCheckoutViewModel.swift
@@ -46,8 +46,6 @@ internal class MercadoPagoCheckoutViewModel: NSObject, NSCopying {
     var hookService: HookService = HookService()
 
     private var advancedConfig: PXAdvancedConfiguration = PXAdvancedConfiguration()
-
-    var paymentResultScreenPreference = PaymentResultScreenPreference()
    
     static var paymentCallback: ((Payment) -> Void)?
     static var finishFlowCallback: ((Payment?) -> Void)?
@@ -295,7 +293,7 @@ internal class MercadoPagoCheckoutViewModel: NSObject, NSCopying {
     }
 
     func resultViewModel() -> PXResultViewModel {
-        return PXResultViewModel(amountHelper: self.amountHelper, paymentResult: self.paymentResult!, instructionsInfo: self.instructionsInfo, paymentResultScreenPreference: self.paymentResultScreenPreference)
+        return PXResultViewModel(amountHelper: self.amountHelper, paymentResult: self.paymentResult!, instructionsInfo: self.instructionsInfo, resultConfiguration: self.advancedConfig.paymentResultConfiguration)
     }
 
     //SEARCH_PAYMENT_METHODS
